refactor: refactor catalogpage
