+++ frontend/src/components/Layout/Layout.tsx
@@ -7,11 +7,11 @@ const Layout: React.FC = () => {
   return (
     <>
       <Header />
-      {/* <div> */}
+
       <Suspense fallback={<div>Loading...</div>}>
         <Outlet />
       </Suspense>
-      {/* </div> */}
+
       <Footer />
     </>
   );
