+++ frontend/src/components/Layout/index.ts
@@ -1,1 +1,2 @@
-export { default } from "./Layout";
+export { default as Layout } from "./Layout";
+export { default as PageLayout } from "./PageLayout";
