+++ frontend/src/components/PageHeading/PageHeading.tsx
@@ -1,0 +1,5 @@
+const PageHeading = () => {
+  return <></>;
+};
+
+export default PageHeading;
