+++ frontend/src/pages/CatalogPage/CatalogPage.tsx
@@ -1,21 +1,14 @@
-import { Pagination, Sort } from "components";
+import { Sort } from "components";
 import { useBooks } from "components/Book";
-import { BreadCrumbs, Label } from "pages/CommonPages.styled";
 import React, { useEffect, useState } from "react";
 import { Outlet, useParams, useSearchParams } from "react-router-dom";
-import {
-  FlexWrapper,
-  PageWrapper,
-  StyledCommonWrapper,
-  Wrapper,
-} from "styles/CommonStyled";
 import ContentSection from "./ContentSection/ContentSection";
+import { PageLayout } from "components/Layout";
 
 const CatalogPage: React.FC = () => {
   const { category, subcategory, link } = useParams();
   const [searchParams, setSearchParams] = useSearchParams();
-  const { books, currentPage, setCurrentPage, totalPages, fetchBooks } =
-    useBooks();
+  const { books, currentPage, setCurrentPage, fetchBooks } = useBooks();
   const [sortBy, setSortBy] = useState("");
   const [orderSort, setOrderSort] = useState("asc");
 
@@ -23,13 +16,11 @@ const CatalogPage: React.FC = () => {
     const page = searchParams.get("page")
       ? Number(searchParams.get("page"))
       : 1;
-    const sortBy = searchParams.get("sortBy") || "";
-    const sortOrder = searchParams.get("orderSort") || "asc";
     const loadData = async () => {
       if (page !== currentPage) {
         setCurrentPage(page);
       }
-      await fetchBooks(category, subcategory, link, page, sortBy, sortOrder);
+      await fetchBooks(category, subcategory, link, page, sortBy, orderSort);
     };
 
     loadData();
@@ -39,14 +30,11 @@ const CatalogPage: React.FC = () => {
     link,
     searchParams,
     currentPage,
+    orderSort,
+    sortBy,
     setCurrentPage,
     fetchBooks,
   ]);
-
-  const handlePageChange = (page: number) => {
-    setCurrentPage(page);
-    setSearchParams({ page: page.toString(), sortBy, orderSort });
-  };
 
   const handleSortChange = (sortKey: string, sortOrder: string) => {
     setSortBy(sortKey);
@@ -55,34 +43,11 @@ const CatalogPage: React.FC = () => {
   };
 
   return (
-    <StyledCommonWrapper>
-      <PageWrapper>
-        <Wrapper>
-          <BreadCrumbs>Каталог | {category} </BreadCrumbs>
-          <Label> {link || subcategory || category || "Усі книги"} </Label>
+    <PageLayout>
+      {!!books.length && <Sort onSortChange={handleSortChange} />}
 
-          <FlexWrapper
-            style={{
-              justifyContent: "center",
-              flexWrap: "wrap",
-              gap: "2rem",
-              position: "relative",
-            }}
-          >
-            {!!books.length && <Sort onSortChange={handleSortChange} />}
-
-            {<Outlet context={{ books }} /> || <ContentSection data={books} />}
-          </FlexWrapper>
-          {!!books.length && (
-            <Pagination
-              currentPage={currentPage}
-              totalPages={totalPages}
-              onPageChange={handlePageChange}
-            />
-          )}
-        </Wrapper>
-      </PageWrapper>
-    </StyledCommonWrapper>
+      {<Outlet context={{ books }} /> || <ContentSection data={books} />}
+    </PageLayout>
   );
 };
