+++ frontend/src/pages/index.ts
@@ -6,5 +6,3 @@ export { default as CatalogPage } from "./CatalogPage/CatalogPage";
 export { default as ContactsPage } from "./ContactsPage/ContactsPage";
 export { default as BookClubPage } from "./BookClubPage/BookClubPage";
 export { default as PrivacyPolicyPage } from "./PrivacyPolicyPage/PrivacyPolicyPage";
-// export { default as FavoritesPage } from "./FavoritePage/FavoritePage";
-// export { default as SearchPage } from "./SearchPage/SearchPage";
